feat: add unit testing CircleCI config
