+++ __tests__/ItemListPage.test.jsx
@@ -2,14 +2,14 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import '@testing-library/jest-dom/extend-expect';
-import ItemListPage from './ItemListPage';
+import ItemListPage from 'components/Items/ItemListPage';
 
 const item = {
   _id: '2',
   user_uid: '323234',
   c_user_uid: '3234',
   photo: '#1112',
-  category: 'Nuts',
+  category: ['Nuts'],
   description: 'Bag of Cashew Nuts',
   expiry: '24/10/2020',
   postcode: 'SR5 4TQ',
