+++ __tests__/MyItemListing.test.jsx
@@ -2,17 +2,17 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import '@testing-library/jest-dom/extend-expect';
-import ItemViewPage from 'components/Items/ItemViewPage';
+import MyItemListing from 'components/Items/MyItemListing';
 
 const item = {
   itemId: '2',
   photoId: '#1112',
   provideUserId: '1',
   collectUserId: '3',
-  category: 'Nuts',
+  category: ['Nuts'],
   description: 'Bag of Cashew Nuts',
   expiryDate: '24/10/2020',
-  location: 'SR5 4ZQ',
+  location: 'SR5 4TQ',
   preferredProvideStartTime: '24/10/2020 10:30AM',
   preferredProvideEndTime: '24/10/2020 5:50PM',
   preferredCollectStartTime: null,
@@ -20,9 +20,11 @@ const item = {
   reservedItem: 'false',
 };
 
-test('renders ItemViewPage Component', () => {
-  const { getByText } = render(<ItemViewPage item={item} />);
-  const linkLocation = getByText(/Location/i);
+test('renders MyItemListing Component', () => {
+  const { getByText } = render(
+    <MyItemListing key={item.ItemId} type="provide" myitem={item} />
+  );
 
-  expect(linkLocation).toBeInTheDocument();
+  const linkExpires = getByText(/Expires/i);
+  expect(linkExpires).toBeInTheDocument();
 });
