+++ package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
+    "test:unit": "react-scripts test --watchAll=false",
     "eject": "react-scripts eject",
     "lint": "eslint ./src --ext .js,.jsx",
     "install:clean": "rm -rf node_modules/ && rm -rf package-lock.json && npm install && npm start",
